+++ vmbot/utils.py
@@ -12,17 +12,7 @@ import requests
 
 from .config import config as vmc
 from .helpers.files import STATICDATA_DB, CACHE_DB
-
-
-class ISK(float):
-    """Represent ISK values."""
-    def __format__(self, format_spec):
-        """Format the ISK value with commonly used prefixes."""
-        for unit in ['', 'k', 'm', 'b']:
-            if self < 1000:
-                return "{}{}".format(format(self, format_spec), unit)
-            self /= 1000
-        return "{}t".format(format(self, format_spec))
+from .helpers.types import ISK
 
 
 class PriceError(Exception):
